ecoScripts/service_now/producer/Image/requirements.txt
tabulate==0.8.3
pykafka
kafka-python
pyyaml
requests
bs4
